refactor: useNewTerminalMenuItems
