+++ src/components/organisms/BottomPaneManager/BottomPaneManager.tsx
@@ -17,10 +17,11 @@ import {Icon} from '@atoms';
 
 import {PreviewSavedCommand, SaveCommand} from '@components/molecules/CommandPreview';
 
+import {useNewTerminalMenuItems} from '@hooks/menuItemsHooks';
+
 import {TerminalType} from '@shared/models/terminal';
 
 import TerminalPane from '../TerminalPane';
-import NewTerminalOptions from '../TerminalPane/NewTerminalOptions';
 import TerminalOptions from '../TerminalPane/TerminalOptions';
 import * as S from './BottomPaneManager.styled';
 
@@ -38,6 +39,8 @@ const BottomPaneManager: React.FC = () => {
 
   const [bottomPaneManagerRef, {height}] = useMeasure<HTMLDivElement>();
   const [tabsContainerRef, {height: tabsContainerHeight}] = useMeasure<HTMLDivElement>();
+
+  const {items: newTerminalMenuItems, onClick: onAddNewTerminalHandler} = useNewTerminalMenuItems();
 
   const onAddTerminalHandler = () => {
     const newTerminalId = uuidv4();
@@ -131,7 +134,11 @@ const BottomPaneManager: React.FC = () => {
               <S.PlusCircleFilled onClick={onAddTerminalHandler} />
             </Tooltip>
 
-            <Dropdown overlay={<NewTerminalOptions />} placement="bottomLeft" trigger={['click']}>
+            <Dropdown
+              menu={{items: newTerminalMenuItems, onClick: menuItem => onAddNewTerminalHandler(menuItem.key)}}
+              placement="bottomLeft"
+              trigger={['click']}
+            >
               <S.DownOutlined />
             </Dropdown>
           </S.NewTabActions>
