+++ src/hooks/menuItemsHooks/index.ts
@@ -1,2 +1,3 @@
 export {useHelpMenuItems} from './useHelpMenuItems';
+export {useNewTerminalMenuItems} from './useNewTerminalMenuItems';
 export {useRefDropdownMenuItems} from './useRefDropdownMenuItems';
